+++ app/src/main/java/org/torproject/android/backup/ZipIt.java
@@ -1,4 +1,4 @@
-package org.torproject.android.zip;
+package org.torproject.android.backup;
 
 
 import java.io.BufferedInputStream;
